+++ RockPaperScissors/RockPaperScissors/Enum/ChoiceMenu.swift
@@ -1,0 +1,13 @@
+//
+//  ChoiceMenu.swift
+//  RockPaperScissors
+//
+//  Created by 권태호 on 28/11/2023.
+//
+
+import Foundation
+
+
+enum ChoiceMenu: String, CaseIterable {
+    case scissor = "1", rock = "2", paper = "3", exit = "0"
+}
