+++ RockPaperScissors/RockPaperScissors/Enum/Result.swift
@@ -1,0 +1,18 @@
+//
+//  Result.swift
+//  RockPaperScissors
+//
+//  Created by 권태호 on 28/11/2023.
+//
+
+import Foundation
+
+
+enum Result: String {
+    case win = "이겼습니다."
+    case lose = "졌습니다."
+    case draw = "비겼습니다."
+    case errorDefault = "잘못된 입력입니다. 다시 시도해주세요."
+    case quitGame = "게임종료"
+}
+
