+++ RockPaperScissors/RockPaperScissors/main.swift
@@ -1,20 +1,10 @@
 import Foundation
 
-enum ChoiceMenu: String, CaseIterable {
-    case scissor = "1", rock = "2", paper = "3", exit = "0"
-}
 
-enum Result: String {
-    case win = "이겼습니다."
-    case lose = "졌습니다."
-    case draw = "비겼습니다."
-    case errorDefault = "잘못된 입력입니다. 다시 시도해주세요."
-    case quitGame = "게임종료"
-}
 
 func endGame() {
-    print("게임 종료")
     isWorking = false
+    print("게임종료")
 }
 
 func displayStartMessage() {
@@ -25,6 +15,8 @@ func displayStartMessage() {
 func calculateResult(comInput: ChoiceMenu, userInput: ChoiceMenu) -> Result {
     if userInput == comInput {
         return .draw
+    } else if userInput == .exit {
+        return .quitGame
     } else if (comInput == .scissor && userInput == .rock) ||
                 (comInput == .rock && userInput == .paper) ||
                 (comInput == .paper && userInput == .scissor) {
@@ -40,16 +32,22 @@ while isWorking {
     displayStartMessage()
     
     if let userInput = readLine(), let userPick = ChoiceMenu(rawValue: userInput) {
-        let comPick = ChoiceMenu.allCases.randomElement() ?? .scissor
+        guard let comPick = ChoiceMenu.allCases.filter({ $0 != .exit }).randomElement() else { continue }
         
         //결과값 받는 메서드 추가
         let result: Result = calculateResult(comInput: comPick, userInput: userPick)
         print(result.rawValue)
         
-        if result == .quitGame || result == .win || result == .lose {
+        
+        if result == .win || result == .lose {
             endGame()
+        } else if result == .quitGame {
+            isWorking = false
+        } else if result == .draw {
+            continue
+        } else {
+            print("잘못된 입력입니다. 다시 시도해주세요.")
         }
-    } else {
-        print("잘못된 입력입니다. 다시 시도해주세요.")
     }
 }
+
